+++ nvflare/private/fed/server/fed_server.py
@@ -43,7 +43,7 @@ from nvflare.private.fed.server.server_runner import ServerRunner
 from nvflare.private.fed.utils.fed_utils import shareable_to_modeldata
 from nvflare.private.fed.utils.messageproto import message_to_proto, proto_to_message
 from nvflare.private.fed.utils.numproto import proto_to_bytes
-from nvflare.widgets.fed_event_runner import ServerFedEventRunner
+from nvflare.widgets.fed_event import ServerFedEventRunner
 
 from .client_manager import ClientManager
 from .run_manager import RunManager
